% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndata.R
\name{syndata}
\alias{syndata}
\title{syndata}
\usage{
syndata(
  df = NULL,
  n = 300,
  cov.mat = NULL,
  cor.mat = NULL,
  means = NULL,
  var.names = NULL,
  var.class = NULL,
  ...
)
}
\arguments{
\item{df}{Data Frame}

\item{n}{Sample Size}

\item{cov.mat}{Covariance matrix}

\item{cor.mat}{Correlation matrix}

\item{means}{Means}

\item{var.names}{Variable Names}

\item{var.class}{Variable Class}

\item{...}{Additional Stuff}
}
\value{
Generates multivariate normal data from either a data frame \code{df} or covariance matrix (\code{cov.mat}) or correlation matrix (\code{cov.mat}) of length \code{n}
}
\description{
syndata
}
