% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndata.R
\name{syndata}
\alias{syndata}
\title{syndata}
\usage{
syndata(df = NULL, n = 300, cov.mat = NULL, means = NULL, varnames = NULL, ...)
}
\arguments{
\item{df}{Data Frame}

\item{n}{Sample Size}

\item{cov.mat}{Covariance matrix}

\item{means}{Means}

\item{varnames}{Variable Names}

\item{...}{Additional Stuff}
}
\value{
Generates multivariate normal data from a either a data frame or covariance matrix (\code{cov.mat}) of length \code{n}
}
\description{
syndata
}
